split alias and function builtins
