+++ src/builtins_vars.c
@@ -16,14 +16,6 @@
 extern int last_status;
 extern int func_return;
 
-struct alias_entry {
-    char *name;
-    char *value;
-    struct alias_entry *next;
-};
-
-static struct alias_entry *aliases = NULL;
-
 struct var_entry {
     char *name;
     char *value;
@@ -32,157 +24,7 @@ struct var_entry {
 
 static struct var_entry *shell_vars = NULL;
 
-struct func_entry {
-    char *name;
-    char *text;
-    Command *body;
-    struct func_entry *next;
-};
 
-static struct func_entry *functions = NULL;
-
-static const char *funcfile_path(void) {
-    const char *env = getenv("VUSH_FUNCFILE");
-    if (env && *env)
-        return env;
-    const char *home = getenv("HOME");
-    if (!home)
-        return NULL;
-    static char path[PATH_MAX];
-    snprintf(path, sizeof(path), "%s/.vush_funcs", home);
-    return path;
-}
-
-static void set_alias(const char *name, const char *value);
-
-static const char *aliasfile_path(void) {
-    const char *env = getenv("VUSH_ALIASFILE");
-    if (env && *env)
-        return env;
-    const char *home = getenv("HOME");
-    if (!home)
-        return NULL;
-    static char path[PATH_MAX];
-    snprintf(path, sizeof(path), "%s/.vush_aliases", home);
-    return path;
-}
-
-static void save_aliases(void) {
-    const char *path = aliasfile_path();
-    if (!path)
-        return;
-    FILE *f = fopen(path, "w");
-    if (!f)
-        return;
-    for (struct alias_entry *a = aliases; a; a = a->next)
-        fprintf(f, "%s=%s\n", a->name, a->value);
-    fclose(f);
-}
-
-static void save_functions(void) {
-    const char *path = funcfile_path();
-    if (!path)
-        return;
-    FILE *f = fopen(path, "w");
-    if (!f)
-        return;
-    for (struct func_entry *fn = functions; fn; fn = fn->next)
-        fprintf(f, "%s() { %s }\n", fn->name, fn->text);
-    fclose(f);
-}
-
-void load_aliases(void) {
-    const char *path = aliasfile_path();
-    if (!path)
-        return;
-    FILE *f = fopen(path, "r");
-    if (!f)
-        return;
-    char line[MAX_LINE];
-    while (fgets(line, sizeof(line), f)) {
-        size_t len = strlen(line);
-        if (len && line[len - 1] == '\n')
-            line[len - 1] = '\0';
-        char *eq = strchr(line, '=');
-        if (!eq)
-            continue;
-        *eq = '\0';
-        set_alias(line, eq + 1);
-    }
-    fclose(f);
-}
-
-void load_functions(void) {
-    const char *path = funcfile_path();
-    if (!path)
-        return;
-    FILE *f = fopen(path, "r");
-    if (!f)
-        return;
-    char line[MAX_LINE];
-    while (fgets(line, sizeof(line), f)) {
-        size_t len = strlen(line);
-        if (len && line[len - 1] == '\n')
-            line[len - 1] = '\0';
-        Command *cmds = parse_line(line);
-        for (Command *c = cmds; c; c = c->next) {
-            if (c->type == CMD_FUNCDEF) {
-                define_function(c->var, c->body, c->text);
-                c->body = NULL;
-            }
-        }
-        free_commands(cmds);
-    }
-    fclose(f);
-}
-
-const char *get_alias(const char *name) {
-    for (struct alias_entry *a = aliases; a; a = a->next) {
-        if (strcmp(a->name, name) == 0)
-            return a->value;
-    }
-    return NULL;
-}
-
-static void set_alias(const char *name, const char *value) {
-    for (struct alias_entry *a = aliases; a; a = a->next) {
-        if (strcmp(a->name, name) == 0) {
-            free(a->value);
-            a->value = strdup(value);
-            return;
-        }
-    }
-    struct alias_entry *new_alias = malloc(sizeof(struct alias_entry));
-    if (!new_alias) {
-        perror("malloc");
-        return;
-    }
-    new_alias->name = strdup(name);
-    new_alias->value = strdup(value);
-    new_alias->next = aliases;
-    aliases = new_alias;
-}
-
-static void remove_alias(const char *name) {
-    struct alias_entry *prev = NULL;
-    for (struct alias_entry *a = aliases; a; prev = a, a = a->next) {
-        if (strcmp(a->name, name) == 0) {
-            if (prev)
-                prev->next = a->next;
-            else
-                aliases = a->next;
-            free(a->name);
-            free(a->value);
-            free(a);
-            return;
-        }
-    }
-}
-
-static void list_aliases(void) {
-    for (struct alias_entry *a = aliases; a; a = a->next)
-        printf("%s='%s'\n", a->name, a->value);
-}
 
 const char *get_shell_var(const char *name) {
     for (struct var_entry *v = shell_vars; v; v = v->next) {
@@ -234,108 +76,6 @@ void free_shell_vars(void) {
         v = n;
     }
     shell_vars = NULL;
-}
-
-void define_function(const char *name, Command *body, const char *text) {
-    for (struct func_entry *f = functions; f; f = f->next) {
-        if (strcmp(f->name, name) == 0) {
-            free(f->name);
-            free(f->text);
-            free_commands(f->body);
-            f->name = strdup(name);
-            f->text = strdup(text);
-            f->body = body;
-            return;
-        }
-    }
-    struct func_entry *fn = malloc(sizeof(struct func_entry));
-    if (!fn) {
-        perror("malloc");
-        return;
-    }
-    fn->name = strdup(name);
-    fn->text = strdup(text);
-    fn->body = body;
-    fn->next = functions;
-    functions = fn;
-}
-
-Command *get_function(const char *name) {
-    for (struct func_entry *f = functions; f; f = f->next) {
-        if (strcmp(f->name, name) == 0)
-            return f->body;
-    }
-    return NULL;
-}
-
-static void free_function_entries(void) {
-    struct func_entry *f = functions;
-    while (f) {
-        struct func_entry *next = f->next;
-        free(f->name);
-        free(f->text);
-        free_commands(f->body);
-        free(f);
-        f = next;
-    }
-    functions = NULL;
-}
-
-void free_functions(void) {
-    save_functions();
-    free_function_entries();
-}
-
-void free_aliases(void) {
-    save_aliases();
-    struct alias_entry *a = aliases;
-    while (a) {
-        struct alias_entry *next = a->next;
-        free(a->name);
-        free(a->value);
-        free(a);
-        a = next;
-    }
-    aliases = NULL;
-}
-
-int builtin_alias(char **args) {
-    if (!args[1]) {
-        list_aliases();
-        return 1;
-    }
-    for (int i = 1; args[i]; i++) {
-        char *eq = strchr(args[i], '=');
-        if (!eq) {
-            fprintf(stderr, "usage: alias name=value\n");
-            continue;
-        }
-        *eq = '\0';
-        set_alias(args[i], eq + 1);
-        *eq = '=';
-    }
-    save_aliases();
-    return 1;
-}
-
-int builtin_unalias(char **args) {
-    if (!args[1]) {
-        fprintf(stderr, "usage: unalias name\n");
-        return 1;
-    }
-    for (int i = 1; args[i]; i++)
-        remove_alias(args[i]);
-    save_aliases();
-    return 1;
-}
-
-int builtin_return(char **args) {
-    int status = 0;
-    if (args[1])
-        status = atoi(args[1]);
-    last_status = status;
-    func_return = 1;
-    return 1;
 }
 
 int builtin_shift(char **args) {
